Reactive Marble Proposal
This commit contains a proposal for an alternate way to test reactive
functions.
It introduces a `Marbles` module to create from string-representations a
sequence of recorded test notifications that can be used as input for a
Hot or Cold observable.

I'm planning to add some documentation about this; but for now the
included tests could do.