Replace custom IObservable implementations with Rx native implementations
Rx guarantees serialization of notifications through scheduling, safeguarding observers, and correct pipeline tear-down behavior. These are implemented in Rx.Net as a set of Producers and Sinks, and they have non-trivial implementations, even for methods which produce a single notification.

It is difficult to correctly implement the IObservable interface, and building up a solution with reactive combinators is always recommended over rolling your own.

This might be a breaking change for code which relied on this incorrect behavior.